Game analyze by AI
Allow players to request a game analyze when game has ended.
The analyze show whether red or blue has the advantage for each move in the game,
and show what was the best move if player did not played it.

The game data are sent to hex distributed ai api,
which will process every move, and send it back.
Analyze is attached to the game, and printed in game sidebar.