Time control, make possible to recreate in past dates
This was not possible to recreate a time control for moves playedAt dates because of time increments,
it was causing time control to elapse because current date was in future when pushing.

Chrono now have a new 'now' parameter that can be set to null or same as 'date' to prevent elapsing.
time controls should now be recreated by passing null in .push(..., null), except for last move
if game is still currently playing.